User Groups [ie. social] Updated.
Testing the administration of user groups, so that we can have social user groups be part of projects.  This is going to allow us to create internal and/or external teams of registered users, and give them access to projects.  (or for that matter any other system we choose to in the future).

Please note this is in addition to adding companies to projects, so we now have the best of both worlds, and rest assured it will be easy, shmeasy to handle. Because the project add form is three fields.  Project Name, Company, and Group.   And thats all it is to create a project.