New model (used) added, and associated with projects.
Now we have a thing called teams (which are really just a display only naming of user_groups) so that you can add whole groups to a project at a time.

So the process would be... create a group/team -- you would probably do this once while setting up your project management system -- for example a team called Developers.  You can then select a team to be in on a project, along with adding companies to a project.