Usable Behavior Created.
You can now make any model add members, or watchers, or whatever you need to call a HABTM relationship to the users table.

In order to do this you simply need to make the  variable set to something like,  = array('Users.Usable' => array('defaultRole' => 'member')), and make the model belongTo Users.UserGroup as you would need to in order to pick groups of users to add to a model relationship.

In the future we will update the behavior so that it checks for that UserGroup relationship, and/or checks the data for an array of users to add instead of a user group to add users from.