Prepared Message model for model/foreign_key relationship.
We are going to make the message model so that messages can be added to any object as well.  This is in anticipation of changing the discussion threads that go on through out the site into a single table where any record can have a discussion about it.   It will serve a private messaging system, a project task discussion board, a conversation about a contact and so on.