Updated Messages with Comments.
Turned messages into a commentable object. This has the effect of making threaded replies all related to a single message.  And since comments are threadable themselves you can basically have an entire forum around a single message.

To test go to /messages/ pick a message and add a comment to it by clicking ADD COMMENT.