IMPORTANT : Override Redirect.
You can now over ride the default redirect specified in the controller by adding a form field which looks like one of the following...

data[Success][redirect]
data[Error][redirect]
data[Override][redirect]

Unless you're using Override.redirect which will override any redirect regardless of whether the controller action was successful or not, you also have to set the second parameter of the redirect() function in your controllers to [success] or [error].  Here's an example...

->redirect('/myurl/', 'success');

If data[Success][redirect] and redirect(, ), where  = success, then the form will control the redirect instead of the controller.  Oh, this is of coure if redirect is called in the controller.  If redirect isn't called, then of course we can't over ride it.