Big ol chunk of code pulled from app_controller.
About 100 lines of code was pulled out of app_controller having to do with getting the viewPath, and was replaced with two lines in the bootstrap files which does the same thing.

Does everything except automatically allow the userRole name as a possible viewPath directory.  But I figured that it hasn't been needed yet, and its possible for userRoles to literally be any name at all so it was sensible to make userRoles a possible viewPath.   I haven't quite come up with another solution for the use case where we might need a certain thing to show up for admins, and a totally separate view for non-admins, but so far its feeling like we can handle it with permission checks in the Html Helper, and Form Helper.  (where the link itself and the form element itself actually checks for whether the logged in user should see it.)

Its really nice to remove that much code though, and replace it with two lines.  Very elegant.