Creating a Contact Tested.
The adding of a contact from the project, add link through to registration of the contact as a user has been tested.  (Note : this leaves the permissions to see the project still, so its not done.)

There is still quite a bit of info that needs to get displayed on contact pages, but you can easily add a company, then add people to that company, then register those people as users, THEN create a project and add that company to it. Then all of those reigistered users will be added to the project.

Note : this is not retroactive.  So users who register afterwards, or are added to the company would not be added as a project user.  So that would need to be coded at some point.