New Format for Plugins Relating to other Plugin Established.
Okay so its become clear that with the model/foreign_key setup we're using, getting one plugin to work with another one was not well defined and/or useful.  But that should be better now with projects as an example.

Within the projects view folder, you'll find a messages.ctp file, a people.ctp file, and coming up a tasks.ctp file and a milestones.ctp file.  (which of course means there are related functions in the controller)

This allows acl access to each tab of the projects plugin, and allows you to have control over how projects interacts with messages, comments, tasks, etc. Even though each of those are separate systems into themselves.

This projects plugin is our template for all other systems moving forward, so please study it and code other plugins in the same way.