Messages now emails all recipients.
When you create a message all users that the message is going to (as long as they have an email address specified in the users table) get an email with the same subject, and the same message, (no truncating or anything) with a link to the messages index so that they can click the message and reply.

Please note : I added  = array(Usable) to messages.  So now you will only see messages that were sent to you.

So if you don't include the client on the message they will not be able to see it when they log in either.  So it will be in effect a private message.