Admin Prefix Removed from Timesheets, and Manual Time Adding Tested.
timesheets are now available without the admin prefix, and make use of the template by user role functionality.  Also tested the ability to manually add time to a task and/or project so that I could update the system with the time that was tracked manually over the last week or so and produce timesheets.