Comment on Individual Tasks.
The workflow for tasks works like this... Make an easy one line or less description for a name, and then add a comment if you need to explain anything further than that one line.  If you get a task assigned and you don't have all the info you should be able to click the task and read the comments and reply in a threaded format all through the project management system.