Zuha Desktop Projects Tie In Re-tested.
You can now track time to tasks, and should.  You will only see projects and tasks that are specifically assigned to you.  If anyone wants to work on making it so that you can mark a task as complete using Zuha Desktop let me know, and I will send you the code for Zuha Desktop.