Did some work on making prices easier on the tech side.
Now the price is in the form that you use to add an item to the cart.  Instead of being calculated after you add to cart.  We have to display the price so it was MUCH easier to just put the price into the form for the OrderItem.  (this pertains mostly to advanced pricing matrix, for pricing by user role of items)