MAJOR Upgrade for Contacts.
Totally redid the contacts database schema an got rid of a bunch of unnecessary stuff.  Its way more simple than it was before, and now you no longer have three database tables for a single contact.  Got rid of the contacts_people, and contacts_companies tables and now they're both combined into a single contacts table with a new is_company field to let us know if the contact is a company or not.

The upside is that contains, and relationships are a lot easier.  The downside is that we have companies and people combined into a single table, and so its not as future proof as it was before.  But the future is too far off, and we need a working system, that works for today, not for tomorrow.