Added New Setting for Resize or Crop of Image Uploads.
Now when you upload a gallery you have the options at /admin/settings to decide whether those images should be resized, or cropped to the right size.  Resize makes for better retention whats in the picture.  Crop will remove some of the picture in exchange for making it the right size.

This is somewhat better because now you have the choice.  But who says more choices are better, its kind of just the lazy way, because I couldn't readily think of a better way to solve both problems.