an update to modal editor to fix the dialog problems.
I'm pretty sure this will not work with some themes still, because of z-index conflicts, but maybe we need to make some kind of theme rule where we don't use anything over 98 on the z-index scale.  Or we need to use javascript to find the highest current z-index and use +1 as the new z-index for our editor. Either way another update will be needed, but for the sites I've tested on, the editor is working now.