Loggable Behavior Added.
You can now log activities for output into a list of activities by simply making your model  = array('Loggable' => array());  Where the settings are how to map fields from your record being saved to the activities model, so that they can be pulled out in a feed.

I'm not over joyed by this, because of the duplication of data, but it was the only fast way around rss feeds needing authentication.  (Meaning we would need to code in a full api with authentication keys in order to use feeds to display non-duplicated data in the same format.