Fixed Category Editing Bug.
There was a bug in the edit page for categories which was causing the model to change to CategoryItem.  It is now a hidden field where the value is whatever the category was when it was created.   Maybe at some point in the future we allow the model to be edited, just didn't have time to put it in today.