Added SSL Support for the Checkout Page.
You can now set a new setting defined as __ORDERS_SSL and make it so that you are forced to https on the checkout page.  You also use the same constant to define some trust logos that appear on the checkout page.  It is nothing more than an html holder though, so that you can copy and paste any trust logo html into the element.