Invoice and Usable Update
The invoice explanation will come later, but the Usable behavior has been updated to automatically handle contact parents if your model belongsTo contact.  By adding contact_id to your form, you can now automatically allow all registered users of that contact to have access to the record.  This was done because we used it in two places -- invoices and projects -- and so instead of copying and pasting very similar functionality, we put it into a single function abstracted to handle both.