Updated Settings MAJOR UPDATE
This update changes the way settings are used by not using the enumerations table at all for them.  The sql is very important here, because it updates the settings table to correct it.

The purpose of this update was so that we aren't pulling setting types from two different places.  Previous to this the types were in the enumerations table, and we had types again in the Setting model.  Now all the types are in the Setting model defined in a function as an array.