Tested User Access to Invoices
This was a second test of the Usable Behavior which turn out perfectly.  Works very well.  The only changes were to small navigation items to get rid of some admin references and added an invoice element for what to show on the invoice index when there are no invoices to show.