Generate Invoices from Projects, and Page Actions in Index Scaffold.
You can now generate an index from a project and all of the times will automatically be added to the invoice line by line.  To test go to /invoices/invoices/generate.

Also added the ability to specify page actions on the scaffold index so that you can have something besides the default Add New Item action.  This is good because now you have a link for looking at different versions of the index, like for example if you can specify the type of invoice you want to see in the index -- ie. paid, unpaid, partpaid -- from the url.  To test go to /invoices/ and click the link Paid Invoices, which will go to /invoices/invoices/index/paid