Invoice Status Automatically Set
There are three options for an invoice status.  paid, partpaid, unpaid.  These statuses are automatically set based on the values given for Invoice Total, and Invoice Balance.  This is good because there is no need to use a drop down to select the status, and thus an extra step for invoice editing is removed, but still used.

To test edit and invoice.