Fixed Login Meta, and Added User Role View Functionality.
You can now specify an admin_prefix for every user role, and when someone from that user role logs in, it will update the user so that they have access to the prefix view associated with that role.

For example, for the administrators user role, we associate that user role with the prefix admin.  When an admin user logs in we update the user so that their prefix is admin.  So you have prefixes on a per user basis, but gets updated on a per user role basis on login.  (if we ever find a use for per user themes we can do it, but haven't found one yet).  The purpose of this is to allow multiple views without adding multiple view files. We would only create a view file if that specific user role needs a different view.