IMPORTANT : Updated Login, so that it actually uses the users.login() function in the UsersController
This could potentially cause problems, because it breaks the normal Auth->login functionality which was specified in the app_controller.   This was done so that the loginMeta() function will work, but again, it could cause other problems I'm not aware of.

If anyone knows a better way of handling a database write during the login function using Auth, then please let us know.