Unless there is a missed file renaming is done.
Please note that when you are only capitalizing a letter and not changing anything else in a file name or folder you sometimes do not get the rename to work.  git has a config setting to deny case changes by default, which you can change.   So, unless there is a file that was missed because of this renaming issue all file renaming is done.

IMPORTANT : If you see a file that needs to be renamed by nothing more than a case change, then you should name the file something like this commit.  Where you change the name to XXXX.bk.php then commit and push, then rename it to what you want it to be and commit and push again.