Added a line required for Cake2.0 to use the isAuthorized function in AppController.
Before this, we weren't getting record level checks because isAuthorized() in AppController wasn't being used.  Now it is used, because we added the line ->Auth->authorize = array('Controller');  which tells Auth to use the controller method for authorizing access to a page.

Note: This information was available in the cake 2.0 book.