Combobox / Autocomplete Form Input Added.
You can now change a select input to a combination box by simply adding two attributes to the Form->input() function.  By adding class=>combobox, and if you want the name of the field to change when a match occurs with a select option then you would add the option ifmatchname => data[Model][field].

A combination box is a select which is a text input if no select is good.  This allows us for example, to see if a contact name already exists when entering a new contact.

The current use case is when adding a contact person you start typing a name, and if that person already exists it will fill in the name input with matching values, and can be seen at /contacts/contacts/add/person/xxxxx (where xxxxx equals a company id that you are relating the person to).