Updated the adding of contact people to companies.
You can now add existing people to existing companies much easier.  (previous to this it would have been a manual db update, or the person you're adding to a company could not have already existed).   Now you simply go to the company, and click -- add person to [company name] -- and you will have an combobox - autocomplete field for adding existing people to a company or adding a new person to a company from a single input.