Fixing Various Bugs to Help Launch a Site.
These bugs mostly had to do with general usage of the pages, and getting some functions to work like this did before the Cake 2.0 upgrade.  Most notably though is that 'page not found' errors still do not use the right template, and I am at a loss for how to make that happen, as the CakeError no longer uses the AppController beforeRender() nor anything from the AppController actually.