IMPORTANT UPDATE PLEASE READ
When you make a site template please make sure that you always wrap template tags in a div, and leave just the content (with as few declared classes and id's as possible in the content) in the included page.

For example it is much much much better to do <div id=header>{page: X}</div>

Than it is to do {page: X},  where Webpage X content = <div id=header>content</div>

For the end user this makes it much cleaner when making changes (they don't need to worry about looking at the source, and making sure the id header is still there).  And it makes the editor layovers work much better. (the yellow boxes that show when you turn on edit mode.)