IMPORTANT Privileges Update - You can Sync Through the UI now.
With this update you can now go to /privileges/sections/ and click [Run Aco Sync], and the acos table will be updated to include all actions from all controllers, and read out the plugins list of updates as it happens.

Made it so that we only run one plugin at a time for the aco sync, using sessions, and then used javascript so that the page would reload until complete. Because the page is reloading, it also reads out the session of updateds acos as you watch.

This is good because now you don't have to use the acl_extras plugin and command line in order to get new features to show up in the privileges management page.