UsableBehavior Update for Private and Public Access.
You can now use this behavior to control both public records (everyone has access) and individual records.  IMPORTANT NOTE, It is different than ACL because it is not for user role access, it is only for individual users having access to individual records.

In order to have a Usable controlled object be public all you have to do is make sure that the object does not have ANY records at all in the Used table.  If it has one record in the used table, then no one has access but the users specified in the used table.  If there is no reference to the object in the used table then everyone has access, its public.