UsableBehavior Redirects When All Results Were Restricted.
Previous to this, if you knew the url to type in you could get some kind of rendering to occur even if the record being called was not supposed to be accessible to the user making the call.

Now we run a quick count to see if the find would have returned any results as is, and if it would have AND its empty afterwards then we know that the UsableBehavior has removed ALL results that would have otherwise been returned.  We then redirect the user to /users/users/restricted.

This is good because now you cannot reach a page that you have not explicitly been given access to AND which is using the UsableBehavior AND has a record in the UsableBehavior controlling it.   (IMPORTANT NOTE : This will not effect models which use the UsableBehavior but the particular record being called has no records in the used table.  That is part of the all used, or not used at all update pushed earlier.)