Media Plugin Upgraded (Joel Please Read Thoroughly).
Added the ability to upload files by simply relating your model to Media in a hasOne or hasMany relationship, and then including a file field in your form with the name... Media.filename, and using saveAll() instead of save() as we should be anyway to save data records.

For Joel... Great work on the media / encodable behavior stuff so far.  There has just been some confusion about what each one is for.  Media should handle the files (all files), including the moving of files around, and validating, etc.   The encodable behavior should be just for encoding and returning a value to the media plugin when its used.

I have updated the beforeSave() so that encodable is attached on the fly, and your behavior should be used by default so that nothing gets broke as is.  But we can make all of the code more clear by keeping functions in neat little boxes of classes.  NOTE : I haven't tested a media upload on the server, so I'm not sure its 100% right yet, but hopefully by looking at my changes in the Media.php model file you can get an idea of what I mean.