Tested Installer on the Server - it works.
You can now install a base zuha site by simply visiting any current zuha site and going to /install.  Besides InstallController you can see what files are added and modified in this commit.

Please note, adding a site on the server will not add it to our git repo of course. But if we would later like to work on this server created site locally then we would login to the server and using the git version installed on the server, commit the new site to the git repo and push to beanstalk so that we can then pull and work on it locally.