Users index does not show the current user anymore.
I think it is pretty reasonable to remove the user looking at the users index from the list of users, since if you want to get to your user account you would typically click your username or profile image and wouldn't need to find it in the users index.  I can only think of minor reasons why you would need it in the users index, so if anyone can think of a great reason it should be there let me know, and I'll find another way to do it.

Ideas... if we did some bulk editing or something, and had a checkbox next to all users, maybe then we would want to include ourselves in that editing?  Just thinking out loud.