Cleaned up the linking in /privileges/
You will now be asked to clear the session before being able to sync the acos table.  This will prevent an error of some kind of session naming conflict or whatever was causing it to fail, when I didn't clear the session first.  You will now click the only action on the page, and then be taken to a explanatory page which also clears your session, and has a button to run the sync'ing action.