Updated Admin Navigation.
Went through and did some reworking of the grey bar that appears at the top when you're logged in as the admin.  It now matches the menu the quick nav in the mobile admin version, except this one has drop downs instead of having to go to the dashboard to see the sub links for sections.