Add to Cart from Index Pages added.
You can now add items to the cart directly from index pages with multiple catalog items on it.

NOTE : If the item has children, it will still only show the view button, and not the add to cart form.  That is because if it has children then you need to choose options for that item, and so you would choose options by viewing the item.

This is good because the number of required clicks has been reduced to checkout.