May have finally made it so that users don't get their user_role screwed up in the aros table.
Previously whenever you would save a user you had to make sure that user_role_id was part of the data array being saved.  If it wasn't then that user would get their aros record messed up, and would not be able to access anything.

Now it should be corrected via the parentNode() function in the UserModel update.