Workflow and Conditions are very finicky.
I worked for hours trying to figure out why workflows events weren't happening correctly, and without changing any substantial code it seemed to just start working again.  I believe, though I don't know for sure that it happened because I emptied the workflow_events table and got rid of the events that weren't triggered.

It may be that the runQueue when running doesn't play nice with un triggered events and might make the wrong workflow event to fire.  ie. an old workflow event fires for a new action that was taken.  Its hard to tell, but hopefully its fixed with it being a bunch of workflow events that won't be there under normal use.