HtmlHelper Image Upgrade for Dynamic Resizing.
You can now use ->Html->image('file', options, extOptions); and the image will be output to the browser as a dynamically resized version of the file if extOptions and a width and height are set.  NOTE : extOptions is a new attribute added to the image() function.

Here is the minimum required in order for images to be dynamically resized.

1. extOptions must at least equal array()
2. options must have a width and height set

3. the image must come from the sites Theme directory

this->Html->image('/theme/default/images/some_image.jpg', array('width' => '100', 'height' => '100'), array());

There are options you can set in the last attribute - extOptions can equal something like...

array('conversion' => 'resize', 'quality' => 100)

the conversion options are : resize, resizeCrop, crop

This upgrade is good because it was done in anticipation of stopping the dynamic resizing during upload and instead doing it during the call of image galleries.  Thereby letting the user change the settings later without needing to re-upload all of the images they have already had resized during upload.