Gallery AppModel and GalleryImage Model Upgrade.
Did a lot of work on how galleries get their settings so that they are handled in the GalleryAppModel and are available to both the GalleryModel, and GalleryImageModel, and added to the data array of Galleries in the afterFind() function of the GalleryAppModel, so that they're available to all views as well.

This is good because now settings are handled in one place and available wherever galleries are used.