Upgraded bootstrap and global files to include Classes.
This update along with the one before it changes the bootstrap and global functions so that they are now within a class.  This was done so that the names are more reflective of what they're doing, where they are and how they relate to the core cake utility files.

So we'll now see things like ZuhaInflector::pricify() instead of just priceFormat(), and ZuhaSet::array_intersect_r() instead of just myIntersect().  From this we can see that it is an update to the cake Inflector, and cake Set classes, and will therefore have a better idea of what the functions are for just by reading the names and cake documentation.