New Element -- List
This element allows you to make a list of any model records for output using the {element: list.X} tag.  Where X is the instance of the list element you want to show. (remember you can have multiple versions of the same element on the same page, so an instance allows you to have different settings for each instance of the element call)

__ELEMENT_LIST_X is the constant and it has a number of configuration variables that you can set.  So between adding that setting, and putting the template tag where you want it to show, we can show a list of records, and control the name fields, show a thumb nail image, the description fields, and some more.

This should be helpful with quickly putting up widgets like, Latest Posts, Latest Users, Most Popular Posts, and so on.  It will probably need a few more configuration variables in the future, but it has a nice start.