New Auto Paginator Filter (and various fixes).
Did a lot of various fixes but the big improvement here is that in addition to the named parameter 'filter' on pagination pages, I added the named parameter 'start' for auto index filtering.

It performs a query of LIKE value % on the specified field.

An example /enumerations/start:type:P/  Will find all of the enumerations which have a Enumeration.type = p %;