Created a much better way of working with enumerations... FINALLY.
So now the idea is to use a hybrid between the enumerations table and hard coded values.  The way to do it is to make sure that we don't have any models that belongsTo() Enumeration.  Instead of the enumeration_id field would instead just be enumeration, or status, or type, etc.  of the type varchar(100) instead of int(10).   Then if there are system values that are needed we do an array merge in the model between the enumerations table and whatever the hard coded system values are.  (LOOK AT : OrderItem::stasuses() for an example).

That allows us to have hard coded values that we set in our controllers and models, but still allow customizations by allowing additional enumeration types.

In the finished example available at /catalogs/catalogs/dashboard you'll see that paid, and shipped are system status types for orders, but you can add new status types at /enumerations/add with the type ORDER_TRANSACTION_STATUS.