Moved auto pagination filtering.
Had to overwrite the core paginate() function so that our auto paginator filtering would only fire when paginate() is called.  We had some errors where it would fire on view pages because of the archive redirect and how paginate var is available to all controller functions.

FIXED, and working awesomely.