Debug Level 3
Added a place where the errors will only show if you set debugger to level 3 or higher.  This allows us to have error pages and custom error pages while debug mode is set to 2.  But there are some cases where the alias redirect causes it to be hard to find the error in question.  If that is the case then try setting debugger=3 and at that point alias redirects will be turned off and regular system errors will show.