Added form input clearing to system.js.
IMPORTANT : If anyone can think of a good way to keep the form inputs from clearing when they are filled with an actual value instead of a label value it would be very appreciated to post a comment on this update in beanstalk.

NOTE : I fixed it so that the User.password field would always be a password field unless the value is Password.