Important Update, Please read all the way.  Install Upgrade.
We are now really close to allowing users to sign up for zuha instantly. You can now visit anydomain.zuha.com, and if I hadn't made the form hidden you could have entered an accessible database info and instantly had a website setup.  The form is hidden because we need to figure out how to handle new databases for new sites.  But in addition to this, there is a change in how we install new sites.

Instead of going to www2.razorit.com/install/site you now need to setup the domain first and go straight to the domain you want to setup.  So we can go to newdomain.com and if it is point to our zuha install you will be asked to fill out the form and install. Same thing for any domain you set up.  If its pointing to a zuha install and doesn't exist you can create it.