Added a New Model for Handling Connections between User Accounts and 3rd party websites.
This model has a very general field called 'value', which is meant to be a serialized array of the info return during authorization of a user to have their account connected to a third party like twitter or facebook.  It simply saves the user_id, the type (ie. facebook, twitter, etc.) and then puts all of the auth values into a single serialized array.