Connected Blog Posting to Twitter.
If the person doing the posting is connected to a twitter account, it will update that person's twitter account with the title and a link to the blog post.
To connect you have to go to /twitter/twitter/connect

To post go to /blogs/blog_posts/add/X where X is the blog id you want to post to.