Converted Contact Detail Type to Varchar instead of _id and removed the direct relationship to the enumerations.
This goes along with the need to get rid of all enumeration relationships, and use varchar fields instead of belongsTo enumerations by some_field_id.  We still have full functionality of being able to use the enumerations table and adding custom types/labels with it, but we don't need to use that confusing relationship.