Completed the new upgrade script.
When you visit /admin it now runs a compare between the schema files and the latest database structure.  If they don't match, then it shows what upgrades need to take place.

This is a big improvement because we can now centralize our db updates better by putting them in the schema files.  Instead of the version folder which will be deleted completely.

When a db update is made, the schema file for that db table needs to be updated, and besides the normal baking of the schema files we have a particular header that is needed in the schema files AND IMPORTANT...

IMPORTANT : If a db column is renamed (as opposed to dropped or added) then it requires that we use the ZuhaSchema::rename function, like it is currently used in the blogs plugin schema file