Only back up tables which have data to back up now.
Instead of making a copy of all tables before we update them we now only make a copy of them if there is data in the table.  This is good because it will keep the number of tables in the database to a minimum.

Note: backed up tables are now prefixed with zbk_, so an example of a table back up is zbk_blog_posts.

Note: backed up tables are deleted on the next update, so if you update two times you will lose the original data. This is so that we cause name conflicts and/or need to have 100 versions of backed up tables.  So we only support one level of backup.