Contacts Index Much Improved
You can now show contacts on the index by type and very quickly toggle which are visible and very quickly filter items further by typing the begginning or part of a name.  It is mobile (iPad) ready and is an extremely fast way to find leads, customers, companies, and people.