Added an Index Search and New Search Type
You can now use a search element which will search whatever index it is on.  You get full control over the inputs from the variables in the element function, and there is a new search type called contains which searchs for LIKE %term% in opposition to LIKE term% which start uses.