#20 Major Update to Privileges.
The usability and look and feel of the privileges plugin has been vastly improved.  It now consists of what in effect is a singe page, and some section synching pages.  Instead of trying to find the section then the action, you now see all sections and actions at once and can scroll through rather easily or use the side navigation to find what you're looking for.

This is great, because it shows how to handle difficult to admin pages.  Also includes an awesome progress bar that shows when you're syncing sections.  Check it out by going to /admin, then clicking Privileges and then click Update Sections & Actions