Upgraded Template Changing
You can now update a template from any page when logged in as the admin, via the admin nav bar.

This update creates a number of new functions for handling the update of template settings, and makes it easier to change the settings and have it sync with other templates.  Just turn on design mode in the bottom right when logged in as admin, then choose a template from the drop down menu and hit save.

You will instantly see the current page with a new template.