Added the prefix variable to alias.
You can now have a prefix for aliases built into the element.  This is good for example if you want all blog posts to have a permanent url of blog/something-long.  In that example the blog part can be set so that all blog posts get it by default.

Note it would still be customizable on a post by post basis, and wouldn't even have to have all blog posts under the virtual blog directory.