It is now possible to make a form more secure by simply adding the option secure = true to the options parameter of the From->create() method.
If you add the option secure equals true, then the next time the form is loaded we add that form's action to a list of secured form actions in the settings.ini file.  Forms that are in this setting require a FormKey to come with the form data. But all of this will be handled by adding the secure option as well.