Extensively tested install
After all of the changes made with the build portion of the install, had to go through and make sure that installation still worked.  It did NOT.  There is one change in here, that may cause problems, but not sure.  It was commented in the OptimizableBehavior, so if alias saving experiences an issue that is where to look.