Pages are auto created when you build a menu link now.
Depending on some semi-complex conditions, a page will be created when you create a menu item.  The page title can't exist already is the main one, and the WebpageMenuItem.page_type (which is not a real db field) needs to be set to content or section.

Also added a default icon for templates which don't have the icon field set.