Creating Sections Dynamically from the Menu in Install Build.
You can now create a section (in addition to the content pages that was just done previous to this).  Creating a section automatically creates two sample sub items, and two sample sub menu links, in addition to the main section and the main link to the section.  All three pages are filled with placeholder content, and the section page shows the sub pages at the bottom of the section.