page: some-element-name, now uses a file instead of the db version
Not sure how much of an enhancement this is, seeing that we still have to pull the id from the db so that it can be edited with the inline editor.  That sql is a lot lighter than the current non file call, but that is because we don't support children elements and so we don't have to use any contains.  I think it's an improvement, but they might be pretty close, at least we have them both so we can benchmark it on a site in the future.