Fixes ACL & adds User belongsTo User (as Self)
Before this set of fixes, users were able to edit other users.
I had to put in a quick check to ensure that the array being
passed to Acl->check() / Acl->permission() were always [0]-based.
I also changed the loop that is used in permission() to check if
the users id is the value for the Acl field.  It did not seem
right before.
Please do a quick test on one of your sites to be sure this fix is good.